dev to prod for the Milestone1 issues completed

+++ public_html/Project/profile.php
@@ -14,22 +14,25 @@ if (isset($_POST["save"])) {
     $stmt = $db->prepare("UPDATE Users set email = :email, username = :username where id = :id");
     try {
         $stmt->execute($params);
+        flash("Username/email updated", "success");
     } catch (Exception $e) {
         if ($e->errorInfo[1] === 1062) {
-            //https://www.php.net/manual/en/function.preg-match.php
+            //https://www.php.net/manual/en/function.preg-match.php                                                       **User will be able to view their profile**
             preg_match("/Users.(\w+)/", $e->errorInfo[2], $matches);
             if (isset($matches[1])) {
                 flash("The chosen " . $matches[1] . " is not available.", "warning");
             } else {
                 //TODO come up with a nice error message
-                echo "<pre>" . var_export($e->errorInfo, true) . "</pre>";
+                //echo "<pre>" . var_export($e->errorInfo, true) . "</pre>";
+                flash("An unexpected error occurred, please try again", "danger");
             }
         } else {
             //TODO come up with a nice error message
-            echo "<pre>" . var_export($e->errorInfo, true) . "</pre>";
+            //echo "<pre>" . var_export($e->errorInfo, true) . "</pre>";
+            flash("An unexpected error occurred, please try again", "danger");
         }
     }
-    //select fresh data from table
+    //select fresh data from table                                                                                       **User will also be able to update their profile**
     $stmt = $db->prepare("SELECT id, email, IFNULL(username, email) as `username` from Users where id = :id LIMIT 1");
     try {
         $stmt->execute([":id" => get_user_id()]);
@@ -43,7 +46,7 @@ if (isset($_POST["save"])) {
         }
     } catch (Exception $e) {
         flash("An unexpected error occurred, please try again", "danger");
-        //echo "<pre>" . var_export($e->errorInfo, true) . "</pre>";
+        echo "<pre>" . var_export($e->errorInfo, true) . "</pre>";
     }
 
 
@@ -51,33 +54,46 @@ if (isset($_POST["save"])) {
     $current_password = se($_POST, "currentPassword", null, false);
     $new_password = se($_POST, "newPassword", null, false);
     $confirm_password = se($_POST, "confirmPassword", null, false);
-    if (isset($current_password) && isset($new_password) && isset($confirm_password)) {
-        if ($new_password === $confirm_password) {
-            //TODO validate current
-            $stmt = $db->prepare("SELECT password from Users where id = :id");
-            try {
-                $stmt->execute([":id" => get_user_id()]);
-                $result = $stmt->fetch(PDO::FETCH_ASSOC);
-                if (isset($result["password"])) {
-                    if (password_verify($current_password, $result["password"])) {
-                        $query = "UPDATE Users set password = :password where id = :id";
-                        $stmt = $db->prepare($query);
-                        $stmt->execute([
-                            ":id" => get_user_id(),
-                            ":password" => password_hash($new_password, PASSWORD_BCRYPT)
-                        ]);
 
-                        flash("Password reset", "success");
-                    } else {
-                        flash("Current password is invalid", "warning");
+    if (strlen($current_password) > 2 && strlen($new_password) > 2 && strlen($confirm_password) > 2)
+    {
+        if (isset($current_password) && isset($new_password) && isset($confirm_password)) {
+            if ($new_password === $confirm_password) {
+                //TODO validate current
+                $stmt = $db->prepare("SELECT password from Users where id = :id");
+                try {
+                    $stmt->execute([":id" => get_user_id()]);
+                    $result = $stmt->fetch(PDO::FETCH_ASSOC);
+                    if (isset($result["password"])) {
+                        if (password_verify($current_password, $result["password"])) {
+                            $query = "UPDATE Users set password = :password where id = :id";
+                            $stmt = $db->prepare($query);
+                            $stmt->execute([
+                                ":id" => get_user_id(),
+                                ":password" => password_hash($new_password, PASSWORD_BCRYPT)
+                            ]);
+
+                            flash("Password reset", "success");
+                        } else {
+                            flash("Current password is invalid", "warning");
+                        }
                     }
+                } catch (Exception $e) {
+                    echo "<pre>" . var_export($e->errorInfo, true) . "</pre>";
+                    flash("An unexpected error occurred, please try again", "danger");
                 }
-            } catch (Exception $e) {
-                echo "<pre>" . var_export($e->errorInfo, true) . "</pre>";
+            } else {
+                flash("New passwords don't match", "warning");
             }
-        } else {
-            flash("New passwords don't match", "warning");
         }
+        else
+        {
+            flash("Unable to reset password at the time. Please try again later.", "warning");
+        }
+    }
+    else
+    {
+        flash("No changes to password were saved.", "warning");
     }
 }
 ?>
@@ -118,6 +134,8 @@ if ($id > -1 /*&& $id !== get_user_id()*/) {
         <a href="?id=<?php se(get_user_id());   ?>" class="btn btn-primary">View Public Profile</a>
         <?php /* Viewing our profile */ ?>
         <form method="POST" onsubmit="return validate(this);">
+            <br>
+            <h5>Email/Username Update</h5>
             <div class="mb-3">
                 <label class="form-label" for="email">Email</label>
                 <input class="form-control" type="email" name="email" id="email" value="<?php se($email); ?>" />
@@ -127,7 +145,9 @@ if ($id > -1 /*&& $id !== get_user_id()*/) {
                 <input class="form-control" type="text" name="username" id="username" value="<?php se($username); ?>" />
             </div>
             <!-- DO NOT PRELOAD PASSWORD -->
-            <div>Password Reset</div>
+            <br>
+            <!--<div--><!--Password Reset</div>-->
+            <h5>Password Reset</h5>
             <div class="mb-3">
                 <label class="form-label" for="cp">Current Password</label>
                 <input class="form-control" type="password" name="currentPassword" id="cp" />
@@ -140,6 +160,7 @@ if ($id > -1 /*&& $id !== get_user_id()*/) {
                 <label class="form-label" for="conp">Confirm Password</label>
                 <input class="form-control" type="password" name="confirmPassword" id="conp" />
             </div>
+            <br>
             <input class="btn btn-primary" type="submit" value="Update Profile" name="save" />
         </form>
     <?php else : ?>
